Revert "chore: refactor"
This reverts commit 9d277a03577e3d6cd24258795587f019e4c29ed3.
