+++ src/core/find-result.js
